WIP: refactor annotation component
