+++ frontend/src/features/campaigns/components/review/TaskModeReview.tsx
@@ -7,8 +7,8 @@ import { useLayoutStore } from '~/features/layout/layout.store';
 import { formatTaskStatus, getTaskStatusColor } from '~/shared/utils/taskStatus';
 import type { TaskStatus } from '~/shared/utils/taskStatus';
 import { extractCentroidFromWKT } from '~/shared/utils/utility';
-import Statistics from '~/features/annotation/components/Statistics';
-import { AnnotationDistributionMap } from '~/features/annotation/components/AnnotationDistributionMap';
+import Statistics from './Statistics';
+import { AnnotationDistributionMap } from './AnnotationDistributionMap';
 import { ExportDropdown } from './ExportDropdown';
 import { Button } from '~/shared/ui/forms';
 import { UserFilterDropdown } from './UserFilterDropdown';
